+++ src/world/entityimpl/entityzombie.c
@@ -1,4 +1,4 @@
 
 #include "entityzombie.h"
 
-void entityzombie_update(uint32_t entity_id, float32_t dt) {}
+void entityzombie_update(uint32_t entity_id, float64_t dt) {}
