+++ src/world/entityimpl/entityzombie.h
@@ -8,7 +8,7 @@
 extern "C" {
 #endif
 
-void entityzombie_update(uint32_t entity_id, float32_t dt);
+void entityzombie_update(uint32_t entity_id, float64_t dt);
 
 #ifdef __cplusplus
 }
