+++ src/world/impl/worldhome.c
@@ -1,4 +1,4 @@
 #include "worldhome.h"
 
 void worldhome_load(world_ctx_t* world_ctx) {}
-void worldhome_update(world_ctx_t* world_ctx, float32_t dt) {}
+void worldhome_update(world_ctx_t* world_ctx, float64_t dt) {}
