+++ src/world/impl/worldhome.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 void worldhome_load(world_ctx_t* world_ctx);
-void worldhome_update(world_ctx_t* world_ctx, float32_t dt);
+void worldhome_update(world_ctx_t* world_ctx, float64_t dt);
 
 #ifdef __cplusplus
 }
