+++ src/world/impl/worldmain.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 void worldmain_load(world_ctx_t* world_ctx);
-void worldmain_update(world_ctx_t* world_ctx, float32_t dt);
+void worldmain_update(world_ctx_t* world_ctx, float64_t dt);
 
 #ifdef __cplusplus
 }
