+++ src/world/player/player.h
@@ -28,7 +28,7 @@ typedef struct {
 player_t* player_create(void);
 void player_destroy(player_t* player);
 void player_setpos(player_t* player, vec2f32_t pos1);
-void player_update(player_t* player, float32_t dt);
+void player_update(player_t* player, float64_t dt);
 player_shared_t* player_shared_devbuf_create(void);
 void player_shared_devbuf_destroy(player_shared_t* devbuf);
 void player_bitfsm_callback(void* playerstate, bitfsm_token_t tok);
